Move reports to reports directory
