+++ src/app/reports/exposed-passwords-report.component.ts
@@ -9,7 +9,7 @@ import { StateService } from "jslib-common/abstractions/state.service";
 import { CipherType } from "jslib-common/enums/cipherType";
 import { CipherView } from "jslib-common/models/view/cipherView";
 
-import { CipherReportComponent } from "./cipher-report.component";
+import { CipherReportComponent } from "../reports/cipher-report.component";
 
 @Component({
   selector: "app-exposed-passwords-report",
