+++ src/app/reports/weak-passwords-report.component.html
